[Refactor] AppDetail 폴더링
