+++ Week2/SwiftUI/Week7/Week7/Models/Banner.swift
